Start timer only after reading package file output
Before, we would trigger the timer callback on a certain interval
without fail; if the time it took to return the package output was
greater than the interval, we would have multiple files open.

Now we start the io watcher, and on its callback stop the io watcher and
start again the ev_timer; when that callback is started we stop the
timer and start the io again.

We also start the io watcher instead of the timer so we get our updates
as soon as possible.