Rename bspwm_info methods to be more consistent
The function names used will be replicated across all info methods:
 - <info>_init
 - <info>_update
 - <info>_free
 - <info>_should_display
 - <info>_print

Unfortunately they will not have the same type signatures; bspwm_update
requires a line of raw text, for instance, and so making macros around
these functions to be more generic may not work.

There is a workaround; we simply copy the line to the bspwm_info
object when updated, and then we only need to call the function on the
struct as it has all the information it needs.